Make test262 pass on Travis CI
The build/test system is still a mix: there are already
test262-related targets in cmake files but the test jobs on Travis
CI still use the hand-written makefile in root. The reason is that
the targets that exist in cmake files only exist when the project
is configured. However, configuration is lost across CI jobs. So,
this is something that should be cleared up later. It just works
for now.