Add Travis CI configuration (#3)
* Add Travis CI configuration

* Remove inaccessible test submodules

* Reactivate test submodules with old but public urls and hashes

* Fix out-of-source-tree build

Ensure that `tidy` target starts looking for the python scripts in
the tools/ subdirectory of the project root.

* Change Travis CI to cmake-based build system

* Make test262 pass on Travis CI

The build/test system is still a mix: there are already
test262-related targets in cmake files but the test jobs on Travis
CI still use the hand-written makefile in root. The reason is that
the targets that exist in cmake files only exist when the project
is configured. However, configuration is lost across CI jobs. So,
this is something that should be cleared up later. It just works
for now.

* Build x86 targets on Travis CI

* Suppress printing compiler invocations during CI build to reduce noise

* Remove explicit invocation of build_third_party.sh during CI build

Let cmake handle that.

* Ensure that libicuNN:i386 is installed when x86 tests are run on CI

* Suppress unused variable warnings during GC build

The build of the GC is full of "'GC_jmp_buf' defined but not used"
warnings. However, as it is a third-party component and is rather
left unchanged, all we can do is suppress those warnings to reduce
the noise.

* Heavy refactoring of .travis.yml

Integrate build and test jobs for each ARCH.MODE target.

This makes caching across jobs unnecessary, thus it makes parallel
CI builds safer.

This also makes the CI more efficient as the cost of setting up
a new virtual machine for every job is reduced (from 25 to 4).

* Fix custom test targets to work with out-of-source-tree builds

And make use of them in Travis CI, i.e., use cmake-generated
makefiles instead of the hand-written global makefile when testing.