Minimize APT dependencies on Travis CI
Only install those packages, which are actually needed. This reduces
job startup/execution time (a bit).

Also update README to highlight which dependencies are mandatory and
which are optional.

Note: the libc++-dev dependency has been removed from Travis CI
config and README alike as it turned out not to be needed by any of
the builds.

Signed-off-by: Akos Kiss <akiss@inf.u-szeged.hu>