Separate GC library into a cmake ExternalProject
This is the start of a build system componentization effort. An
extra benefit of breaking GC into a cmake ExternalProject is that
this moves the build of libgc under the same build directory as the
rest of the project (i.e., it will not pollute the
third_party/GCutil source directory).

Signed-off-by: Akos Kiss <akiss@inf.u-szeged.hu>