Add nodejs detection to SpiderMonkey test runner
Node.js is only available as `nodejs` on some older systems while newer
platforms either symlink that to `node` or don't provide the tool by the
old name at all. (Homebrew-installed Node.js on macOS falls into the
latter category.)

To support all platforms, this commit adds a detection logic to the
SpiderMonkey test runner and uses the command that's available on the
system.

Signed-off-by: Akos Kiss <akiss@inf.u-szeged.hu>