Reorder Travis CI jobs to reduce total run time
The darwin.x64.release job is considerably slower than its linux
counterpart. Moreover, right now the darwin jobs are the last to
execute. The result of this is that most of the jobs have long
finished but the single longest running job (darwin.x64.release)
still has to be waited for.

Signed-off-by: Akos Kiss <akiss@inf.u-szeged.hu>