Make tests run on macOS (#43)
* Change V8 test runner not to compare whole test execution logs

When running V8 tests on macOS, the test runner reported failure
although the logs have clearly shown that all tests passed. It has
turned out that a single line of text was 'missing' from the logs:
"No connection to distribution server; running tests locally."

This has revealed the following problems:
- On macOS (actually, on anything non-Linux), "network distribution"
  is disabled early on and test script doesn't get to check for
  distribution server. So it cannot print that line in the logs
  either.
- Our test framework shouldn't try network distribution at all, not
  even on Linux.
- The outcome of the V8 test execution is determined by comparing
  the whole test execution log to a reference log with no
  difference allowed. However, the reference log cannot be updated
  (at least it's hard as it's in a submodule).

This commit
- adds the `--no-network` flag to the test script to prevent
  network distribution attempts in the future and to make all
  platforms behave the same, and
- replaces the outcome test from a log diff to a 'grep' for an
  "All tests succeeded" message.

Signed-off-by: Akos Kiss <akiss@inf.u-szeged.hu>

* Add nodejs detection to SpiderMonkey test runner

Node.js is only available as `nodejs` on some older systems while newer
platforms either symlink that to `node` or don't provide the tool by the
old name at all. (Homebrew-installed Node.js on macOS falls into the
latter category.)

To support all platforms, this commit adds a detection logic to the
SpiderMonkey test runner and uses the command that's available on the
system.

Signed-off-by: Akos Kiss <akiss@inf.u-szeged.hu>

* Make Travis CI run tests for macOS, too

Note: For now, chakracore and spidermonkey tests are disabled for
macOS as they would need fixes outside this repository.

Signed-off-by: Akos Kiss <akiss@inf.u-szeged.hu>

* Reorder Travis CI jobs to reduce total run time

The darwin.x64.release job is considerably slower than its linux
counterpart. Moreover, right now the darwin jobs are the last to
execute. The result of this is that most of the jobs have long
finished but the single longest running job (darwin.x64.release)
still has to be waited for.

Signed-off-by: Akos Kiss <akiss@inf.u-szeged.hu>