Move to addon-based dependency installation and drop sudo requirement on Travis CI
It especially helps the darwin jobs: on macOS, the dependeny
installation via 'manual' brew invocation takes about 4.5 mintes,
while the Homebrew addon brings that down to cca. 10 seconds.

Addons remove the need for `sudo: required`. But it is not replaced
with `sudo: false` as the whole `sudo` key has been deprecated. So,
it is simply removed.

Signed-off-by: Akos Kiss <akiss@inf.u-szeged.hu>