Link libicu on ARM-Linux devices
ESCARGOT_LIBICU_SUPPORT build option have added:
If turned ON, then enable libicu. It turned ON by default.

In case of Linux host, libicu is only linked to Escargot on x86 and x64 architectures.
This patch helps to link libicu on ARM-Linux devices too, so linker problems can be avoided.

Signed-off-by: Csaba Repasi repasics@inf.u-szeged.hu