Add memory statistics feature to the project.
Added a boehm-gc heap usage monitoring feature to the project.
A runtim flag (--mem-stats) helps to show the collected informaion.

Renamed ESCARGOT_PROFILE_MASSIF to ESCARGOT_MEM_STATS in the CMake.
Of course, Valgrind based tools still can be used. In this case, the
project should be compiled with ESCARGOT_VALGRIND definition.

Signed-off-by: Roland Takacs <rtakacs.uszeged@partner.samsung.com>