Update LexicalEnvironment instead of creating new ExecutionState objects.
The motivation of this patch is that the state is a c++ reference, and cannot
be reassigned in the interpreter. However, recursively calling interpret for
each let/const block would cost too much. This patch is the first path of reworking
with statements to work without recursive calls. The same technique could
be used by let/const blocks later.

Signed-off-by: Zoltan Herczeg zherczeg.u-szeged@partner.samsung.com