Fix arrow function bugs related to lazy instantiation
Fixed a bug when an arrow function with block braces is inside an arrow function without braces handled incorrently during an outer function lazy instantiation.
The inner arrow function treated the most outer simple function as its parent instead of the outer arrow function due to the invalid use of config.parseSingleFunction.
Also the 0 parseFunctionIndex must be valid for arrow functions as well during the lazy instantiation.

This patch fixes #343.

Signed-off-by: Robert Fancsik frobert@inf.u-szeged.hu